Patch for #(3591399) Parser error - Null-Pointer-Exception
When I put fName into the SVDBGenerateIf.java module, there is a case
where an if statement can be recognized as a generate statement and visa
versa.  If fName hasn't been initalized (something to do with the
casting?) we get a null pointer because styled string is trying to style
something that doesn't exist.

I have put in something here which is probably more of a patch than a
fix.  The TODO there should hopefully have enough information in it to
send someone more skilled than I (Matthew?) in the direction of a proper
fix.